Update automation-onboard-solutions.md
The proposed changes are not comprehensive; the section should be rewritten to be clearer. I'm  not sure what the message of the sentence "To onboard Azure Virtual Machines at scale, an existing VM must be onboarded with the Change tracking or Update management solution." is, so further clarification should be made - is this a prerequisite to following this article, or does this article address this prerequisite? Does it mean that the Update Management solution needs to be initialized by a single VM before applying it using automation to many VMs? Or do I need to configure an application inside a management VM to be able to use the solution at scale? 
Additionally, the Update Management and Change Tracking solutions are separate solutions, which makes this paragraph even less understandable. 

There's discussion on Reddit discussing this section. https://www.reddit.com/r/AZURE/comments/858hnw/who_writes_azure_documentation/